+++ App.tsx
@@ -3,6 +3,7 @@
  */
 
 import './src/i18n';
+import './src/config/geolocationInit';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
 import { GestureHandlerRootView } from 'react-native-gesture-handler';
 import { AppNavigationRoot } from './src/AppNavigationRoot';